refactor(optimus): refactor
