+++ optimus/lib/optimus.dart
@@ -31,7 +31,9 @@ export 'src/common/group_item.dart';
 export 'src/common/scroll.dart';
 export 'src/date_time_field.dart';
 export 'src/dialogs/dialog.dart';
+export 'src/dialogs/dialog_size.dart';
 export 'src/dialogs/dialog_wrapper.dart';
+export 'src/dialogs/inline_dialog.dart';
 export 'src/enabled.dart';
 export 'src/expansion/expansion_tile.dart';
 export 'src/form/checkbox_form_field.dart';
