+++ optimus/lib/src/dialogs/dialog.dart
@@ -1,37 +1,6 @@
-import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 import 'package:optimus/optimus.dart';
-import 'package:optimus/src/common/anchored_overlay.dart';
-
-enum OptimusDialogSize {
-  /// Intended for short, to the point messages.
-  ///
-  /// Not suitable for dialogs with a lot of UI elements. Mostly used for
-  /// non-modal dialogs to deliver users a message without preventing them from
-  /// interacting with the application. Footers in small dialogs use only
-  /// vertical button groups.
-  small,
-
-  /// The most common dialog size.
-  ///
-  /// Suitable for any content except those with very complex UI.
-  regular,
-
-  /// Intended for dialogs with complex UI elements (tables, forms with multiple
-  /// columns, etc.) or large components like images.
-  ///
-  /// It can only be used as a modal dialog with centered position.
-  large,
-}
-
-enum OptimusDialogType {
-  /// Default dialog type. Used for common action.
-  common,
-
-  /// Primary button has [OptimusDialogType.destructive] variant, used for
-  /// destructive actions.
-  destructive,
-}
+import 'package:optimus/src/dialogs/dialog_content.dart';
 
 /// The position of a dialog is determined by functionality.
 enum OptimusDialogPosition {
@@ -160,10 +129,13 @@ class OptimusDialog extends StatelessWidget {
 
   final Widget content;
 
+  /// {@template optimus.dialog.wrapper}
   /// Builds custom content. If content padding needed wrap in
   /// [OptimusDialogContentPadding].
+  /// {@endtemplate}
   final ContentWrapperBuilder? contentWrapperBuilder;
 
+  /// {@template optimus.dialog.actions}
   /// Controls dialog actions.
   ///
   /// First button should always contain primary action. When single button it
@@ -177,6 +149,7 @@ class OptimusDialog extends StatelessWidget {
   ///
   /// All other buttons have [OptimusButtonVariant.text] variant and represent
   /// additional actions.
+  /// {@endtemplate}
   final List<OptimusDialogAction> actions;
 
   /// Controls dialog size.
@@ -235,7 +208,7 @@ class OptimusDialog extends StatelessWidget {
 
     return Align(
       alignment: _alignment(context),
-      child: _DialogContent(
+      child: DialogContent(
         title: title,
         content: content,
         actions: actions,
@@ -252,127 +225,6 @@ class OptimusDialog extends StatelessWidget {
   }
 }
 
-class OptimusInlineDialog extends StatelessWidget {
-  const OptimusInlineDialog({
-    Key? key,
-    required this.content,
-    this.contentWrapperBuilder,
-    this.actions = const [],
-    this.close,
-    required this.anchorKey,
-  }) : super(key: key);
-
-  final VoidCallback? close;
-  final GlobalKey anchorKey;
-  final Widget content;
-  final ContentWrapperBuilder? contentWrapperBuilder;
-  final List<OptimusDialogAction> actions;
-
-  @override
-  Widget build(BuildContext context) => AnchoredOverlay(
-        anchorKey: anchorKey,
-        width: OptimusDialogSize.small.width,
-        child: _DialogContent(
-          content: content,
-          actions: actions,
-          type: OptimusDialogType.common,
-          size: OptimusDialogSize.small,
-          maxWidth: OptimusDialogSize.small.width,
-          contentWrapperBuilder: contentWrapperBuilder,
-          isDismissible: true,
-          close: close,
-        ),
-      );
-}
-
-class _DialogContent extends StatelessWidget {
-  const _DialogContent({
-    Key? key,
-    required this.actions,
-    required this.content,
-    required this.type,
-    required this.size,
-    required this.maxWidth,
-    this.title,
-    this.close,
-    this.isDismissible,
-    this.contentWrapperBuilder,
-    this.spacing,
-    this.margin,
-  }) : super(key: key);
-
-  final List<OptimusDialogAction> actions;
-  final Widget? title;
-  final double? spacing;
-  final double maxWidth;
-  final EdgeInsetsGeometry? margin;
-
-  final Widget content;
-  final VoidCallback? close;
-  final OptimusDialogSize size;
-  final bool? isDismissible;
-  final OptimusDialogType type;
-
-  final ContentWrapperBuilder? contentWrapperBuilder;
-
-  Widget _divider(OptimusThemeData theme) =>
-      Divider(height: 1, color: theme.colors.neutral50);
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = OptimusTheme.of(context);
-    final titleValue = title;
-
-    return Container(
-      margin: margin,
-      child: Padding(
-        padding: EdgeInsets.all(spacing ?? 0),
-        child: ConstrainedBox(
-          constraints: BoxConstraints(maxWidth: maxWidth),
-          child: OptimusCard(
-            variant: OptimusBasicCardVariant.overlay,
-            padding: OptimusCardSpacing.spacing0,
-            child: Material(
-              color: theme.isDark
-                  ? theme.colors.neutral500
-                  : theme.colors.neutral0,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                mainAxisSize: MainAxisSize.min,
-                children: <Widget>[
-                  if (titleValue != null)
-                    _Title(
-                      title: titleValue,
-                      close: close ?? () => Navigator.pop(context),
-                      isDismissible: isDismissible ??
-                          ModalRoute.of(context)?.barrierDismissible ??
-                          true,
-                    ),
-                  if (titleValue != null) _divider(theme),
-                  OptimusParagraph(
-                    child: _Content(
-                      content: content,
-                      contentWrapperBuilder: contentWrapperBuilder,
-                    ),
-                  ),
-                  if (actions.isNotEmpty) _divider(theme),
-                  if (actions.isNotEmpty)
-                    _Actions(
-                      actions: actions,
-                      type: type,
-                      dialogSize: size,
-                      close: close ?? () => Navigator.pop(context),
-                    )
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
 class OptimusDialogContentPadding extends StatelessWidget {
   const OptimusDialogContentPadding({
     Key? key,
@@ -384,142 +236,4 @@ class OptimusDialogContentPadding extends StatelessWidget {
   @override
   Widget build(BuildContext context) =>
       Padding(padding: const EdgeInsets.all(spacing200), child: child);
-}
-
-class _Content extends StatelessWidget {
-  const _Content({
-    Key? key,
-    required this.content,
-    this.contentWrapperBuilder,
-  }) : super(key: key);
-
-  final Widget content;
-  final ContentWrapperBuilder? contentWrapperBuilder;
-
-  @override
-  Widget build(BuildContext context) {
-    final contentWrapperBuilder = this.contentWrapperBuilder;
-    final wrappedContent = contentWrapperBuilder == null
-        ? OptimusScrollConfiguration(
-            child: SingleChildScrollView(
-              child: OptimusDialogContentPadding(child: content),
-            ),
-          )
-        : contentWrapperBuilder(context, content);
-
-    return Flexible(fit: FlexFit.loose, child: wrappedContent);
-  }
-}
-
-class _Title extends StatelessWidget {
-  const _Title({
-    Key? key,
-    required this.title,
-    required this.close,
-    required this.isDismissible,
-  }) : super(key: key);
-
-  final Widget title;
-  final VoidCallback close;
-  final bool isDismissible;
-
-  @override
-  Widget build(BuildContext context) => Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Expanded(
-            child: Padding(
-              padding: const EdgeInsets.all(spacing200),
-              child: OptimusSubsectionTitle(child: title),
-            ),
-          ),
-          if (isDismissible)
-            Padding(
-              padding: const EdgeInsets.only(top: spacing50),
-              child: OptimusIconButton(
-                icon: const OptimusIcon(iconData: OptimusIcons.cross_close),
-                variant: OptimusIconButtonVariant.bare,
-                onPressed: close,
-              ),
-            ),
-        ],
-      );
-}
-
-class _Actions extends StatelessWidget {
-  const _Actions({
-    Key? key,
-    required this.actions,
-    required this.dialogSize,
-    required this.type,
-    required this.close,
-  }) : super(key: key);
-
-  final List<OptimusDialogAction> actions;
-  final OptimusDialogSize dialogSize;
-  final OptimusDialogType type;
-  final VoidCallback close;
-
-  bool get _isVertical => dialogSize == OptimusDialogSize.small;
-
-  @override
-  Widget build(BuildContext context) {
-    final children = actions
-        .mapIndexed<Widget>(
-          (i, e) => Padding(
-            padding: EdgeInsets.only(
-              bottom: _isVertical ? spacing200 : 0,
-              left: _isVertical ? 0 : spacing200,
-            ),
-            child: OptimusButton(
-              onPressed: e.onPressed ?? close,
-              minWidth: _isVertical ? double.infinity : null,
-              variant: _getVariant(i),
-              key: e.key,
-              child: e.title,
-            ),
-          ),
-        )
-        .toList();
-    if (children.length > 2 && !_isVertical) {
-      children.insert(2, const Spacer());
-    }
-
-    return Padding(
-      padding: EdgeInsets.only(
-        left: _isVertical ? spacing200 : 0,
-        right: spacing200,
-        top: spacing200,
-        bottom: _isVertical ? 0 : spacing200,
-      ),
-      child: Flex(
-        mainAxisAlignment: MainAxisAlignment.end,
-        direction: _isVertical ? Axis.vertical : Axis.horizontal,
-        children: children.reversed.toList(),
-      ),
-    );
-  }
-
-  OptimusButtonVariant _getVariant(int i) {
-    if (type == OptimusDialogType.destructive && i == 0) {
-      return OptimusButtonVariant.destructive;
-    }
-    if (i == 0) return OptimusButtonVariant.primary;
-    if (i == 1) return OptimusButtonVariant.defaultButton;
-
-    return OptimusButtonVariant.text;
-  }
-}
-
-extension on OptimusDialogSize {
-  double get width {
-    switch (this) {
-      case OptimusDialogSize.small:
-        return 320;
-      case OptimusDialogSize.regular:
-        return 576;
-      case OptimusDialogSize.large:
-        return 896;
-    }
-  }
 }
